restructured the code briefly to make it easier to underestand
