+++ app/lab/chicagoCTA/page.tsx
@@ -1,4 +1,4 @@
-import AllTrainAlerts from '@/src/components/Laboratory/Experiments/ChicagoCTA/AllTrainAlerts';
+import AllTrainAlerts from '@/src/components/pages/Laboratory/Experiments/ChicagoCTA/AllTrainAlerts';
 export default async function ChicagoCTA() {
   return <AllTrainAlerts />;
 }
