+++ src/components/pages/Home/Quotes/Quotes.tsx
