+++ src/components/pages/Laboratory/Experiments/ChicagoCTA/TrainAlertView.tsx
