+++ src/components/reusable/cards/BarebonesCard.tsx
