--- src/mockData/projects.json
@@ -1,95 +1,0 @@
-[
-  {
-    "id": 101,
-    "title": "Build a Birdhouse",
-    "tags": ["woodworking", "DIY", "outdoors"],
-    "featured": true,
-    "description": "A project to construct a wooden birdhouse for your backyard.",
-    "details": [
-      {
-        "order": 1,
-        "id": 1,
-        "title": "Gather Materials",
-        "detail": "Collect all necessary tools and wood for the birdhouse."
-      },
-      {
-        "order": 2,
-        "id": 2,
-        "title": "Cut and Assemble",
-        "detail": "Cut the wood into pieces and assemble the birdhouse structure."
-      },
-      {
-        "order": 3,
-        "id": 3,
-        "title": "Paint and Decorate",
-        "detail": "Paint the birdhouse and add decorative elements."
-      }
-    ],
-    "images": [
-      "https://example.com/images/birdhouse1.jpg",
-      "https://example.com/images/birdhouse2.jpg"
-    ]
-  },
-  {
-    "id": 102,
-    "title": "Learn TypeScript",
-    "tags": ["coding", "typescript", "programming"],
-    "featured": false,
-    "description": "A step-by-step guide to mastering TypeScript for JavaScript developers.",
-    "details": [
-      {
-        "order": 1,
-        "id": 1,
-        "title": "Setup Development Environment",
-        "detail": "Install Node.js and set up a TypeScript compiler."
-      },
-      {
-        "order": 2,
-        "id": 2,
-        "title": "Learn Basic Types",
-        "detail": "Understand and use basic TypeScript types like string, number, and boolean."
-      },
-      {
-        "order": 3,
-        "id": 3,
-        "title": "Advanced Features",
-        "detail": "Learn about interfaces, generics, and advanced TypeScript features."
-      }
-    ],
-    "images": [
-      "https://example.com/images/typescript1.jpg",
-      "https://example.com/images/typescript2.jpg"
-    ]
-  },
-  {
-    "id": 103,
-    "title": "Bake a Cake",
-    "tags": ["baking", "cooking", "desserts"],
-    "featured": true,
-    "description": "A simple recipe to bake a delicious chocolate cake.",
-    "details": [
-      {
-        "order": 1,
-        "id": 1,
-        "title": "Prepare Ingredients",
-        "detail": "Measure and prepare all the ingredients needed for the cake."
-      },
-      {
-        "order": 2,
-        "id": 2,
-        "title": "Mix Batter",
-        "detail": "Combine ingredients to make the cake batter."
-      },
-      {
-        "order": 3,
-        "id": 3,
-        "title": "Bake and Decorate",
-        "detail": "Bake the cake and decorate with frosting and toppings."
-      }
-    ],
-    "images": [
-      "https://example.com/images/cake1.jpg",
-      "https://example.com/images/cake2.jpg"
-    ]
-  }
-]
